Calorie_Counter/requirements.txt
python-telegram-bot
google-generativeai
gspread
oauth2client
python-dotenv
Pillow